Additional code that should be for refactor/tests branch
