+++ app/Http/Controllers/API/PlayerController.php
@@ -180,5 +180,7 @@ class PlayerController extends Controller
         return response()->json($data);
     }
 
+    
+
 
 }
