+++ database/factories/UserFactory.php
@@ -23,12 +23,20 @@ class UserFactory extends Factory
      */
     public function definition(): array
     {
+        do {
+            $fakeName = $this->faker->firstName;
+        } while (strlen($fakeName <5));
+        do {
+            $fakeSurname = $this->faker->lastName;
+        } while (strlen($fakeSurname <5));
+
         return [
-            'name' => fake()->name(),
-            'email' => fake()->unique()->safeEmail(),
+            'name'              => $fakeName,
+            'surname'           => $fakeSurname,
+            'email'             => fake()->unique()->safeEmail(),
             'email_verified_at' => now(),
-            'password' => static::$password ??= Hash::make('password'),
-            'remember_token' => Str::random(10),
+            'password'          => static::$password ??= Hash::make('password'),
+            'remember_token'    => Str::random(10),
         ];
     }
